+++ eelbrain/plot/_uts.py
@@ -12,7 +12,7 @@ from . import _base
 from ._base import (
     EelFigure, PlotData, Layout,
     LegendMixin, YLimMixin, XAxisMixin, TimeSlicerEF, frame_title)
-from ._colors import colors_for_oneway, find_cell_colors
+from ._colors import colors_for_oneway, find_cell_styles
 from .._colorspaces import oneway_colors
 from functools import reduce
 
@@ -123,7 +123,7 @@ class UTSStat(LegendMixin, XAxisMixin, YLimMixin, EelFigure):
                  main=np.mean, error='sem', pool_error=None, legend='upper right', labels=None,
                  axtitle=True, xlabel=True, ylabel=True, xticklabels='bottom',
                  invy=False, bottom=None, top=None, hline=None, xdim='time',
-                 xlim=None, clip=None, color='b', colors=None, error_alpha=0.3,
+                 xlim=None, clip=None, colors=None, error_alpha=0.3,
                  clusters=None, pmax=0.05, ptrend=0.1, *args, **kwargs):
         # coerce input variables
         sub, n = assub(sub, ds, return_n=True)
@@ -168,10 +168,7 @@ class UTSStat(LegendMixin, XAxisMixin, YLimMixin, EelFigure):
             nax = len(ct.cells)
 
         # assemble colors
-        if color_x is None:
-            colors = {None: color}
-        else:
-            colors = find_cell_colors(color_x, colors)
+        styles = find_cell_styles(color_x, colors)
 
         layout = Layout(nax, 2, 4, *args, **kwargs)
         EelFigure.__init__(self, frame_title(y, x, xax), layout)
@@ -182,7 +179,7 @@ class UTSStat(LegendMixin, XAxisMixin, YLimMixin, EelFigure):
         self._plots = []
         legend_handles = {}
         if xax is None:
-            p = _ax_uts_stat(self._axes[0], ct, colors, main, error, dev_data, xdim, hline, clusters, pmax, ptrend, clip, error_alpha)
+            p = _ax_uts_stat(self._axes[0], ct, styles, main, error, dev_data, xdim, hline, clusters, pmax, ptrend, clip, error_alpha)
             self._plots.append(p)
             legend_handles.update(p.legend_handles)
             if len(ct) < 2:
@@ -198,7 +195,7 @@ class UTSStat(LegendMixin, XAxisMixin, YLimMixin, EelFigure):
                     match = matchct.data[cell]
 
                 ct_ = Celltable(ct.data[cell], X_, match=match, coercion=asndvar)
-                p = _ax_uts_stat(ax, ct_, colors, main, error, dev_data, xdim, hline, clusters, pmax, ptrend, clip, error_alpha)
+                p = _ax_uts_stat(ax, ct_, styles, main, error, dev_data, xdim, hline, clusters, pmax, ptrend, clip, error_alpha)
                 self._plots.append(p)
                 legend_handles.update(p.legend_handles)
                 self._set_axtitle(axtitle, names=map(cellname, ct.cells))
@@ -402,7 +399,7 @@ class UTS(TimeSlicerEF, LegendMixin, YLimMixin, XAxisMixin, EelFigure):
 
 class _ax_uts_stat:
 
-    def __init__(self, ax, ct, colors, main, error, dev_data, xdim, hline, clusters, pmax, ptrend, clip, error_alpha):
+    def __init__(self, ax, ct, styles, main, error, dev_data, xdim, hline, clusters, pmax, ptrend, clip, error_alpha):
         # stat plots
         self.ax = ax
         self.stat_plots = []
@@ -412,8 +409,7 @@ class _ax_uts_stat:
         for cell in ct.cells:
             ndvar = ct.data[cell]
             y = ndvar.get_data(('case', xdim))
-            plt = _plt_uts_stat(ax, x, y, main, error, dev_data, colors[cell],
-                                cellname(cell), clip, error_alpha)
+            plt = _plt_uts_stat(ax, x, y, main, error, dev_data, styles[cell], cellname(cell), clip, error_alpha)
             self.stat_plots.append(plt)
             if plt.main is not None:
                 self.legend_handles[cell] = plt.main[0]
@@ -692,16 +688,14 @@ class _plt_uts_clusters:
 
 class _plt_uts_stat:
 
-    def __init__(self, ax, x, y, main, error, dev_data, color, label, clip,
-                 error_alpha):
+    def __init__(self, ax, x, y, main, error, dev_data, style, label, clip, error_alpha):
         # plot main
         if hasattr(main, '__call__'):
             y_main = main(y, axis=0)
             lw = mpl.rcParams['lines.linewidth']
             if error == 'all':
                 lw *= 2
-            self.main = ax.plot(x, y_main, color=color, label=label, lw=lw,
-                                zorder=5, clip_on=clip)
+            self.main = ax.plot(x, y_main, label=label, lw=lw, zorder=5, clip_on=clip, **style.line_args)
         elif error == 'all':
             self.main = None
         else:
@@ -709,8 +703,7 @@ class _plt_uts_stat:
 
         # plot error
         if error == 'all':
-            self.error = ax.plot(x, y.T, color=color, alpha=error_alpha,
-                                 clip_on=clip)
+            self.error = ax.plot(x, y.T, alpha=error_alpha, clip_on=clip, **style.line_args)
         elif error and len(y) > 1:
             if error == 'data':
                 pass
@@ -720,8 +713,6 @@ class _plt_uts_stat:
                 dev_data = stats.variability(y, None, None, error, False)
             lower = y_main - dev_data
             upper = y_main + dev_data
-            self.error = ax.fill_between(x, lower, upper, color=color,
-                                         alpha=error_alpha,
-                                         linewidth=0, zorder=0, clip_on=clip)
+            self.error = ax.fill_between(x, lower, upper, color=style.color, alpha=error_alpha, linewidth=0, zorder=0, clip_on=clip)
         else:
             self.error = None